refactor: api response 핸들링 로직 네이밍 수정
