+++ app/src/main/java/woowacourse/shopping/ShoppingApp.kt
@@ -28,7 +28,6 @@ class ShoppingApp : Application() {
 
     override fun onTerminate() {
         super.onTerminate()
-        productSource.shutDown()
         historyProductDb.close()
     }
