+++ app/src/main/java/woowacourse/shopping/data/order/remote/datasource/OrderRemoteDataSource.kt
@@ -1,11 +1,14 @@
 package woowacourse.shopping.data.order.remote.datasource
 
-import woowacourse.shopping.data.common.ResponseHandlingUtils.handleExecute
+import woowacourse.shopping.data.common.ApiResponseHandler.handleApiResponse
 import woowacourse.shopping.data.common.ResponseResult
 import woowacourse.shopping.data.order.remote.OrderApiService
 import woowacourse.shopping.data.order.remote.dto.OrderRequest
 
 class OrderRemoteDataSource(private val orderApiService: OrderApiService) : OrderDataSource {
     override suspend fun orderCartItems(cartItemIds: List<Long>): ResponseResult<Unit> =
-        handleExecute { orderApiService.createOrder(OrderRequest(cartItemIds)) }
+        handleApiResponse {
+            val orderRequest = OrderRequest(cartItemIds)
+            orderApiService.createOrder(orderRequest)
+        }
 }
