+++ app/src/main/java/woowacourse/shopping/data/payment/remote/datasource/CouponRemoteDataSource.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.payment.remote.datasource
 
-import woowacourse.shopping.data.common.ResponseHandlingUtils.handleExecute
+import woowacourse.shopping.data.common.ApiResponseHandler.handleApiResponse
 import woowacourse.shopping.data.common.ResponseResult
 import woowacourse.shopping.data.payment.remote.CouponApiService
 import woowacourse.shopping.data.payment.remote.dto.CouponResponse
@@ -9,5 +9,5 @@ class CouponRemoteDataSource(
     private val couponApiService: CouponApiService,
 ): CouponDataSource {
     override suspend fun loadCoupons(): ResponseResult<CouponResponse> =
-        handleExecute { couponApiService.requestCoupons() }
+        handleApiResponse { couponApiService.requestCoupons() }
 }
