+++ app/src/main/java/woowacourse/shopping/data/product/remote/datasource/ProductRemoteDataSource.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.product.remote.datasource
 
-import woowacourse.shopping.data.common.ResponseHandlingUtils.handleExecute
+import woowacourse.shopping.data.common.ApiResponseHandler.handleApiResponse
 import woowacourse.shopping.data.common.ResponseResult
 import woowacourse.shopping.data.product.remote.ProductsApiService
 import woowacourse.shopping.data.product.remote.dto.ProductDto
@@ -10,18 +10,13 @@ class ProductRemoteDataSource(
     private val productsApiService: ProductsApiService,
 ) : ProductDataSource {
     override suspend fun loadByPaged(page: Int): ResponseResult<ProductResponse> =
-        handleExecute { productsApiService.requestProducts(page = page) }
+        handleApiResponse { productsApiService.requestProducts(page = page) }
 
     override suspend fun loadById(id: Long): ResponseResult<ProductDto> =
-        handleExecute { productsApiService.requestProduct(id.toInt()) }
+        handleApiResponse { productsApiService.requestProduct(id.toInt()) }
 
     override suspend fun loadByCategory(category: String): ResponseResult<ProductResponse> =
-        handleExecute { productsApiService.requestProducts(category) }
-
-    override fun shutDown(): Boolean {
-        // TODO: 연결 끊기
-        return false
-    }
+        handleApiResponse { productsApiService.requestProducts(category) }
 
     companion object {
         private const val TAG = "ProductRemoteDataSource"
